Add files via upload
Aim of project.
To predict the price of houses with the available independent variables. 
To build a regression model  in order to arrive at a prediction 

This model will then be used by the management to understand how exactly the prices vary with the variables. 
They can accordingly manipulate the strategy of the firm and concentrate on areas that will yield high returns. 
Further, the model will be a good way for management to understand the pricing dynamics of a new market. 

Linear, Ridge and Lasso Regression models used